Extract layout ops to layout_ops.rs (refs #68)
Moves resize_pane_if_needed() and handle_divider_drag() out of main.rs
into a dedicated layout_ops.rs module. Both operate on the PaneTree
layout: the first resizes a PTY when its pane rect changes, the
second handles interactive split-divider dragging.

Mirrors wezterm-gui's termwindow/render/split.rs + resize.rs.

main.rs drops from 475 to 377 LoC.

No behavior change — pure mechanical refactor.
